Mastodon configuration fix; security group fix RE 443
Update the SMTP endpoint and port used by the Rails mailer to deliver email.
Outbound SMTP on 587 appears to be blocked within Scaleway so we can use 2525
which is otherwise OK.

Update the https rule in the security group to use port 443 instead of port 80.
Why was I able to reach the instance while this was broken? Worrisome.